requirements.txt
# ---- 核心数值与数据处理 ----
numpy>=1.21
pandas>=1.4

# ---- 环境变量管理 ----
python-dotenv>=1.0

# ---- OKX 官方 SDK ----
python-okx ==0.4.0
